remove custom css from index.html
specifying width and height to be a percentage of vw for the entire project can be problematic if we ever want the page to scroll

this is hurting us hard in studio-ui. very many things that should be easy to do but is now incredibly cumbersome due to having made assumptions about overflow/width/height on an app-level. more details in https://apollographql.atlassian.net/browse/DEBT-4

the same result can be achieved a different way. will tackle that in the next commit

Signed-off-by: Chang Wang <garnwraly@gmail.com>