make storybooks representative of the pages (#12)
* remove custom css from index.html

specifying width and height to be a percentage of vw for the entire project can be problematic if we ever want the page to scroll

this is hurting us hard in studio-ui. very many things that should be easy to do but is now incredibly cumbersome due to having made assumptions about overflow/width/height on an app-level. more details in https://apollographql.atlassian.net/browse/DEBT-4

the same result can be achieved a different way. will tackle that in the next commit

Signed-off-by: Chang Wang <garnwraly@gmail.com>

* have planet not stretch the page by making localized assertions

we want to show a circle that's larger than the page

we want that circle not to stretch the page

we can achieve that by creating a localized constraint around the circle rather than around the page to preserve the flexibility of the page

Signed-off-by: Chang Wang <garnwraly@gmail.com>

* update storybooks

Signed-off-by: Chang Wang <garnwraly@gmail.com>

* add link to font file

Signed-off-by: Chang Wang <garnwraly@gmail.com>